moving deprecation function into utils
