+++ lomap/gufe_bindings/network_generation.py
@@ -5,49 +5,18 @@ from gufe import(
 )
 import gufe
 import itertools
-import functools
 import logging
 import networkx as nx
 import numpy as np
-from typing import Callable, Optional, Union, Any, Dict
-import warnings
+from typing import Callable, Optional, Union
 
 from ..graphgen import GraphGen
 from .._due import due, Doi
-
+from ..utils import deprecated_kwargs
 
 logger = logging.getLogger(__name__)
 
-def rename_kwargs(func_name:str, kwargs:Dict[str, Any], old_name:str, new_name:str):
-    """Helper function for deprecating function arguments."""
-    if old_name in kwargs:
-        if new_name in kwargs:
-            raise ValueError(f"Both {new_name} and {old_name} are defined for {func_name}. {old_name} is deprecated, please use {new_name} instead.")
-
-        else:
-          warnings.warn(f"{func_name} argument '{old_name}' is deprecated. Please use '{new_name}' instead.",
-                      DeprecationWarning)
-          kwargs[new_name] = kwargs.pop(old_name)
-    return kwargs
-
-def deprecated_kwarg(old_name:str, new_name:str) -> Callable:
-  """Decorator for deprecating keyword arguments
-
-  e.g.
-  @deprecated_kwarg(old_arg, new_arg)
-  def my_function(new_arg)
-      ...
-  """
-  def decorator(func: Callable)->Callable:
-
-      @functools.wraps(func)
-      def wrapper(*args:Any, **kwargs:Any):
-            kwargs = rename_kwargs(func.__name__, kwargs, old_name=old_name, new_name=new_name)
-            return func(*args, **kwargs)
-      return wrapper
-  return decorator
-
-@deprecated_kwarg(old_name='molecules', new_name='ligands')
+@deprecated_kwargs(name_mappings={'molecules':'ligands'})
 @due.dcite(Doi("https://doi.org/10.1007/s10822-013-9678-y"), description="LOMAP")
 def generate_lomap_network(
         ligands: list[gufe.SmallMoleculeComponent],
